Imoproved tools layout calulation in order to avoid empty spaces among the grid
