+++ src/solutions-page/components/tools/Tools.js
@@ -2,29 +2,37 @@ import { EmptyTools } from './EmptyTools';
 import { Tool } from './tool';
 import { useFilter } from '../../contexts/FilterContext';
 import { getTools } from 'common/utils';
-
-const wideCards = [
-	'20085485-7185-40fd-89e4-14dbb690aea2', // Advanced Review
-	'ad68e506-8c2b-4c0f-a9e3-16623d00041e', // Booking & Appointments
-];
+import { useRef, useEffect, useState } from '@wordpress/element';
 
 export const Tools = () => {
+	const grid = useRef();
 	const { category, search } = useFilter();
-	let tools = getTools( { category, search } );
+	const [ tools, setTools ] = useState( getTools( { category, search } ) )
 
-	tools.map( tool => {
-		if ( wideCards.includes( tool.id ) ) {
-			tool.wide = true;
+	const calculateTools = () => {
+		if ( grid?.current ) {
+			const columns = window.getComputedStyle( grid.current )[ 'grid-template-columns' ].split( ' ' ).length
+			setTools( getTools( { category, search, columns } ) )
 		}
+	}
 
-		return tool;
-	} )
+	useEffect( () => {
+		window.addEventListener( 'resize', calculateTools );
+
+		return () => {
+			window.removeEventListener( 'resize', calculateTools );
+		};
+	}, [tools] );
+
+	useEffect( () => {
+		calculateTools();
+	}, [category, search] );
 
 	return (
 		<>
 			{ ! tools.length && <EmptyTools/> }
 			{ !! tools.length && (
-				<div className="nfd-solutions-tools nfd-grid nfd-gap-4 nfd-grid-cols-2 min-[1200px]:nfd-grid-cols-3 min-[1520px]:nfd-grid-cols-4">
+				<div className="nfd-solutions-tools nfd-grid nfd-gap-4 nfd-grid-cols-2 min-[1200px]:nfd-grid-cols-3 min-[1520px]:nfd-grid-cols-4" ref={ grid }>
 					{ tools.map( ( tool ) => (
 						<Tool
 							name={ tool?.name }
